fix(task-service): preserve rubric score precision and add essay feedback null-safety
- Change RubricScoreData.score from Integer to Double to preserve decimal precision
  from AI evaluation scores (e.g., 45.5 instead of 45)
- Update TaskCompletionMapper.extractScoreData() to use Double instead of casting to int
- Add comprehensive null-safety checks to EssayTaskEvaluator.buildSuccessEvent()
- Extract feedback string building into separate method (buildFeedbackString)
  for better maintainability and null-checking
- Add validation for feedback and overall evaluation objects
- Use default values when optional fields are null
- Add warning logs for null evaluation components with submission ID for debugging

+++ skilltracker-common/common-event/src/main/java/com/amalitech/common/event/events/TaskCompletedEvent.java
@@ -76,9 +76,10 @@ public class TaskCompletedEvent {
     @Builder
     public static class RubricScoreData {
         /**
-         * Score achieved for this rubric.
+         * Score achieved for this rubric (as floating-point to preserve precision).
+         * Example: 45.5 for a score of 45.5 points
          */
-        private Integer score;
+        private Double score;
 
         /**
          * Maximum possible score for this rubric.
